Integrate camera prototyp
