+++ default_hometemp.ini
@@ -1,6 +1,6 @@
 [hometemp]
-version = 0.4-DEV
-sensor_pin = 4
+version = v0.4-BaseTemp
+sensor_pin = 17
 
 [db]
 container_name =
