+++ pictures/commanded/.gitkeep
@@ -1,0 +1,1 @@
+default location for pictures (requested via command)
